Prototype: Move SVG jsx components into subfolder
