+++ prototype/src/js/svg/ArrowLeftSVG.jsx
