+++ prototype/src/js/svg/ArrowRightSVG.jsx
