+++ prototype/src/js/svg/ArrowUpSVG.jsx
