+++ prototype/src/js/svg/DeviceSVG.jsx
