+++ prototype/src/js/svg/TargetSVG.jsx
